Enforce mandate jurisdiction/policy and harden money arithmetic
Mandate evaluation parsed jurisdictions and policies but never enforced them
(spec §6.1 steps 6–7). It now refuses an intent whose fulfillment
jurisdiction isn't authorized — treating a listed ISO country (US) as
authorizing its subdivisions (US-CA, the form context.jurisdiction actually
carries) — and honors the prohibit_subscriptions policy.

Money arithmetic:
- The 8.75% tax used a plain i64 multiply while every other amount op
  saturated; a crafted unit price near i64::MAX panicked under
  overflow-checks or wrapped to a corrupt/negative tax that flowed into the
  charged total and receipt. Widened to i128.
- decimal_to_minor now saturates on overflow instead of collapsing to 0 (the
  worst fallback for a charge) and logs it.
- An unrecognized currency code is logged rather than silently persisting the
  order in the engine's default currency.
- Corrected the negotiate discount comment to match its actual truncation.

+++ src/service/helpers.rs
@@ -135,7 +135,13 @@ pub(super) fn price_request_items(
             total: Money::new(line_subtotal, currency),
         });
     }
-    let tax_minor = subtotal_minor * 875 / 10_000;
+    // 8.75% tax. Widen to i128 for the multiply: `subtotal_minor` can reach
+    // i64::MAX (the line math above saturates), and a plain `i64 * 875`
+    // would overflow — panicking under overflow-checks or silently wrapping
+    // to a corrupt (even negative) tax that then flows into the charged
+    // total and the signed receipt. The result is < subtotal, so the
+    // narrowing back to i64 cannot truncate.
+    let tax_minor = (i128::from(subtotal_minor) * 875 / 10_000) as i64;
     let total_minor = subtotal_minor.saturating_add(tax_minor);
     let totals = Totals {
         subtotal: Some(Money::new(subtotal_minor, currency)),
